Fixing import statements (#66)
* cleaning array_processing from old module

* cleaning out unused dependencies

* trivial function from riogrande duplicated - aviding absolute import

* fixing imports convster.io.py

* removing unused imports

* setting imports for convster.parallel.py

* convster.plotting.py (was already) ready

* remove unused imports

* started adapting test_convster.conftest.py - wait for riogrande imports fix

* io.py fixed imports and removed unused

* note for later

* io_.py fixed imports

* first fix of parallel.py imports

* removing unused imports

* done with imports in linfit.inference.py

* parallel.py done

* done imports in linfit.parallel.py

* removing unused imports

* shuffling functions to resolve dependency loop

---------
