+++ src/landiv_blur/__init__.py
@@ -1,5 +1,3 @@
-from . import _array_processing as ap
-
 from .io import (
     load_block,
 )
