+++ src/riogrande/io.py
@@ -2,38 +2,29 @@ from __future__ import annotations
 
 import os
 import glob
-import rasterio
-
 
 from math import floor
 
-from typing import Any
-
+import rasterio # TODO: choose either rio or rasterio as name
 import rasterio as rio
 from rasterio.io import DatasetWriter
-from rasterio.enums import ColorInterp
 from rasterio.windows import Window
-from rasterio.enums import Resampling
 from rasterio.mask import mask
 from rasterio.warp import (
     calculate_default_transform,
     reproject,
     Resampling,
-    transform_bounds
 )
 
 from shapely.geometry import box as shbox
-import geopandas as gpd
 
 from numpy.typing import NDArray
 
-from ._exceptions import (
+from .exceptions import (
     BandSelectionNoMatchError,
     BandSelectionAmbiguousError,
-    SourceNotSavedError,
-    UnknownExtensionError,
 )
-from ._helper import (
+from .helper import (
     check_crs_raster,
     outfile_suffix,
     serialize,
@@ -41,7 +32,6 @@ from ._helper import (
     sanitize,
     match_all,
     view_to_window,
-    get_scale_factor,
 )
